refactor(core): Project architecture refactoring

+++ lib/src/main/kotlin/com/fleetmate/lib/data/dto/car/CarCreateDto.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class CarCreateDto(
     val registrationNumber: String,
-    val fuelLevel: Float,
-    val mileage: Float,
+    val fuelLevel: Double,
+    val mileage: Double,
     val additionalDate: Long,
     val type: Int
 )
