+++ lib/src/main/kotlin/com/fleetmate/lib/data/dto/car/CarUpdateDto.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class CarUpdateDto(
     val registrationNumber: String? = null,
-    val fuelLevel: Float? = null,
-    val mileage: Float? = null,
+    val fuelLevel: Double? = null,
+    val mileage: Double? = null,
     val dateAdded: Long? = null,
     val type: Int? = null
 )
