+++ lib/src/main/kotlin/com/fleetmate/lib/data/dto/report/ReportUpdateDto.kt
@@ -4,8 +4,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ReportUpdateDto(
-    val mileage: Float?,
-    val avgSpeed: Float?,
+    val mileage: Double?,
+    val avgSpeed: Double?,
     val trip: Int?,
     val car: Int?,
     val driver: Int?
