+++ trip-service/src/main/kotlin/com/fleetmate/trip/modules/report/data/dto/ReportFullOutputDto.kt
@@ -1,13 +1,13 @@
 package com.fleetmate.trip.modules.report.data.dto
 
-import com.fleetmate.lib.dto.trip.TripFullOutputDto
+import com.fleetmate.lib.data.dto.trip.TripFullOutputDto
 import kotlinx.serialization.Serializable
 
 @Serializable
 data class ReportFullOutputDto(
     val id: Int?,
-    val mileage: Float?,
-    val avgSpeed: Float?,
+    val mileage: Double?,
+    val avgSpeed: Double?,
     val trip: TripFullOutputDto?
 ) {
 }
